+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [next]
 * BREAKING CHANGE: improvement in animations to SimplePlayer and SimpleEnemy. Now use `SimpleDirectionAnimation` to manipulate animations.
 * add `GameColorFilter` it is now possible to add color filter in the game.
+* Possible to load maps made by Tiled by the server. Just pass the link as path.
 
 ## [0.6.27]
 * little performance improvement;